Refactor context and token management for SearchInput and Tags components
- Updated import paths for useComponentToken to improve clarity and organization.
- Introduced initComponentTokens to centralize token initialization for SEARCH_INPUT and TAGS components.
- Enhanced ThemeContext by removing the old useComponentToken implementation and replacing it with a more structured approach.
- Updated SearchInput and Tags components to utilize the new token management system, ensuring better type safety and maintainability.
- Adjusted TagsDemo to accommodate changes in token structure.
