+++ lib/components/Inputs/SearchInput/SearchInput.tsx
@@ -5,7 +5,7 @@ import PrimitiveInput from "../../Primitives/PrimitiveInput/PrimitiveInput";
 
 import { SearchInputProps } from "./types";
 import { SearchInputTokensType } from "./searchInput.tokens";
-import { useComponentToken } from "../../../context";
+import { useComponentToken } from "../../../context/useContextToken";
 
 const toPixels = (value: string | number | undefined): number => {
   if (typeof value === "number") {
