+++ lib/context/ThemeProvider.tsx
@@ -1,24 +1,11 @@
-import { getSearchInputTokens } from "../components/Inputs/SearchInput/searchInput.tokens";
-import { getTagTokens } from "../components/Tags/tag.tokens";
 import { FOUNDATION_THEME, ThemeType } from "../tokens";
+import initTokens from "./initComponentTokens";
 import ThemeContext, { ComponentTokenType } from "./ThemeContext";
 
 type ThemeProviderProps = {
   foundationTokens?: ThemeType;
   componentTokens?: ComponentTokenType;
   children: React.ReactNode;
-};
-
-const initTokens = (
-  componentTokens: ComponentTokenType,
-  foundationTokens: ThemeType
-): Required<ComponentTokenType> => {
-  return {
-    TAGS: componentTokens.TAGS ?? getTagTokens(foundationTokens),
-    SEARCH_INPUT:
-      componentTokens.SEARCH_INPUT ?? getSearchInputTokens(foundationTokens),
-    // add supprort for other components here
-  };
 };
 
 const ThemeProvider = ({
